Replace usage in dials.index code with new concat function
