+++ tests/array_family/test_flex_shoebox.py
@@ -137,27 +137,3 @@ def test_bounding_boxes():
     bbox2 = shoebox.bounding_boxes()
     for i in range(10):
         assert bbox2[i] == bbox[i]
-
-
-def test_concat():
-    from dials.array_family import flex
-
-    table1 = flex.reflection_table()
-    table2 = flex.reflection_table()
-
-    table1["id"] = flex.size_t([0, 0, 1, 1])
-    table2["id"] = flex.size_t([0, 0, 1, 1])
-
-    ids1 = table1.experiment_identifiers()
-    ids2 = table2.experiment_identifiers()
-
-    ids1[0] = "a"
-    ids1[1] = "b"
-    ids2[0] = "c"
-    ids2[1] = "d"
-
-    table1 = flex.reflection_table.concat([table1, table2])
-
-    assert list(table1["id"]) == [0, 0, 1, 1, 2, 2, 3, 3]
-    assert list(ids1.keys()) == [0, 1, 2, 3]
-    assert list(ids1.values()) == ["a", "b", "c", "d"]
